+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## Version 0.2 (Sept 8, 2019 JST)
+
+- support multiple break points (Destructive change)
+  - remove function: `setBreakPoint`
+  - add functions: `addBreakPoint` , `removeBreakPoint` , `removeAllBreakPoints`
+
 ## Version 0.1 (Sept 7, 2019 JST)
 
 - implemented the all of documented instructions